move get_scale, get_angle, get_sign to frame.py & use get_affine
