+++ scarlet2/measure.py
@@ -4,7 +4,7 @@ import math
 import astropy.units as u
 
 from .source import Component
-
+from .frame import get_scale, get_angle, get_sign
 
 def max_pixel(component):
     """Determine pixel with maximum value
@@ -324,37 +324,6 @@ def binomial(n, k):
         result *= n - i + 1
         result //= i
     return result
-
-def get_scale(wcs):
-    c1 = (wcs.pc[0,:2]**2).sum()**0.5
-    c2 = (wcs.pc[1,:2]**2).sum()**0.5
-    return jnp.array([c1, c2])
-    
-def get_angle(wcs):
-    c = get_scale(wcs)
-    c = c.reshape([c.shape[-1],1])
-    R = wcs.pc[:2, :2] / c # removing the scaling factors from the pc
-
-    if R[0,0]==0.:
-        return jnp.arcsin(R[0,1])
-    else:
-        return jnp.arctan(R[0,1]/R[0,0])
-
-def get_sign(wcs):
-    c = get_scale(wcs)
-    c = c.reshape([c.shape[-1],1])
-    R = wcs.pc[:2, :2] / c # removing the absolute scaling factors from the pc
-
-    if R[0,0]==0.:
-        phi = jnp.arcsin(R[0,1])
-    else:
-        phi = jnp.arctan(R[0,1]/R[0,0])
-    
-    R_inv = jnp.array([[jnp.cos(phi), -jnp.sin(phi)],
-                    [jnp.sin(phi), jnp.cos(phi)]])
-    
-    R = R_inv @ R
-    return jnp.diag(R)
